test: Move tests to correct packages
