+++ json/src/test/java/uk/gov/android/authentication/json/jwk/JWKTest.kt
@@ -1,12 +1,11 @@
-package uk.gov.android.authentication.integrity.appcheck.usecase
+package uk.gov.android.authentication.json.jwk
 
+import org.jose4j.jwk.JsonWebKey
 import java.security.KeyFactory
 import java.security.PublicKey
 import java.security.spec.X509EncodedKeySpec
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import org.jose4j.jwk.JsonWebKey
-import uk.gov.android.authentication.json.jwk.JWK
 
 class JWKTest {
     @Test
@@ -22,7 +21,7 @@ class JWKTest {
             -122, 72, -50, 61, 3, 1, 7, 3, 66, 0, 4, -45, 113, 67, 94, -79, 85, -108, 14, -79, -116,
             0, 101, -101, -60, 32, 60, -20, -128, 70, -65, -62, -54, 113, -13, 53, 14, 39, -18, 56,
             25, -83, 94, -102, -33, -118, 27, -89, 80, -51, 21, -54, -58, -19, -118, -30, -47, 94,
-            36, -57, -114, -17, -52, -94, 125, -85, -110, -54, -50, -18, 17, -14, 82, 1, 32
+            36, -57, -114, -17, -52, -94, 125, -85, -110, -54, -50, -18, 17, -14, 82, 1, 32,
         )
         val key = createMockKey(keyInBytes)
         val actual = JWK.generateJwk(key)
@@ -38,21 +37,21 @@ class JWKTest {
                 "sig",
                 "P-256",
                 X,
-                Y
-            )
+                Y,
+            ),
         )
         private val joseJwk = JsonWebKey.Factory.newJwk(
             mapOf(
                 "kty" to "EC",
                 "x" to "03FDXrFVlA6xjABlm8QgPOyARr_CynHzNQ4n7jgZrV4",
                 "y" to "mt-KG6dQzRXKxu2K4tFeJMeO78yifauSys7uEfJSASA",
-                "crv" to "P-256"
-            )
+                "crv" to "P-256",
+            ),
         )
         private fun createMockKey(key: ByteArray): PublicKey {
             val factory = KeyFactory.getInstance("EC")
             return factory.generatePublic(
-                X509EncodedKeySpec(key)
+                X509EncodedKeySpec(key),
             )
         }
     }
